requirements.txt
fastapi==0.135.1
uvicorn==0.41.0
sqlalchemy==2.0.48
asyncpg==0.31.0
psycopg2-binary==2.9.11
pgvector==0.4.2
alembic==1.18.4
python-dotenv==1.2.2
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
pydantic[email]==2.12.5
python-multipart==0.0.20
httpx==0.28.1
pymupdf4llm==0.0.17
redis==7.3.0
starlette==0.52.1
